introduced fairing and payload adapter sizing
sizing made thanks to the function "buckling" and "mat_switch". now the optimal staging is assessed considering also the fairing separation (together with stage 1) and adapter separation (together with stage 2). The fairing sizing also provides a plot of the lateral view of the fairing, the adapter and the payload. I also revised the struct for M1 and M2: now they both include the section "other" (for any other mass that is not yet sized, like the recovery system for stage 1 or avionics for stage 2, etc etc).